api/organizations: move inmemory db implementation to the db package
Created with Sturdy
